python/requirements.txt
python-dotenv==0.19.2
requests==2.26.0
aiohttp==3.8.1
privateai-client==1.2.0